requirements.txt
blinker==1.9.0
certifi==2026.2.25
click==8.3.2
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pymongo==4.16.0
python-dotenv==1.2.2
Werkzeug==3.1.8
